+++ EliteAPI.Events/Status/Ship/Events/BodyRadiusStatusEvent.cs
@@ -2,11 +2,11 @@
 
 namespace EliteAPI.Events.Status.Ship.Events;
 
-public readonly struct BodyRadiusStatusEvent : IStatusEvent<float>
+public readonly struct BodyRadiusStatusEvent : IStatusEvent<double>
 {
     public DateTime Timestamp => DateTime.Now;
     
     public string Event => "BodyRadius";
     
-    public float Value { get; init; }
+    public double Value { get; init; }
 }
