+++ EliteAPI.Events/Status/Ship/Events/GravityStatusEvent.cs
@@ -2,11 +2,11 @@
 
 namespace EliteAPI.Events.Status.Ship.Events;
 
-public readonly struct GravityStatusEvent : IStatusEvent<float>
+public readonly struct GravityStatusEvent : IStatusEvent<double>
 {
     public DateTime Timestamp => DateTime.Now;
     
     public string Event => "Gravity";
     
-    public float Value { get; init; }
+    public double Value { get; init; }
 }
